requirements.txt
amqp==2.3.2
backcall==0.1.0
billiard==3.5.0.4
celery==4.2.1
decorator==4.3.0
Django>=2.1
django-celery-beat==1.1.1
django-celery-results==1.0.1
django-ckeditor==5.6.1
django-js-asset==1.1.0
django-model-utils==3.1.2
django-notifications-hq==1.5.0
django-registration-redux==2.4
django-widget-tweaks==1.4.2
factory-boy==2.11.1
Faker==0.8.17
ipython==6.5.0
ipython-genutils==0.2.0
jedi==0.12.1
jsonfield==2.0.2
kombu==4.2.1
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.2.0
prompt-toolkit==1.0.15
ptyprocess==0.6.0
Pygments==2.2.0
python-dateutil==2.7.3
pytz==2018.5
redis==2.10.6
simplegeneric==0.8.1
six==1.11.0
text-unidecode==1.2
django-notifications-hq==1.5.0
django-model-utils==3.1.2
traitlets==4.3.2
vine==1.1.4
wcwidth==0.1.7
